requirements.txt
asgiref==3.11.0
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
sqlparse==0.5.4
tzdata==2025.2
psycopg2-binary
python-dateutil